[FEAT] 팔로우 했는지 확인하는 기능 추가 - 엔티티 Long 타입 int 타입으로 변경

+++ src/main/java/com/ssafy/sgdc/follow/Follow.java
@@ -23,7 +23,7 @@ public class Follow {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // auto increment
     @Column(name = "friend_id")
-    private long friendId;
+    private int friendId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
